Fix matchups
