+++ src/types.ts
@@ -1,4 +1,4 @@
-import { Division } from './constants'
+import { Division } from '@/constants'
 
 export type KeyOf<T> = keyof T
 export type ValueOf<T> = T[KeyOf<T>]
