+++ src/utils/translation.ts
@@ -1,5 +1,5 @@
-import { RankMapping } from '../constants'
-import { kanjiToNumber } from './japanese'
+import { RankMapping } from '@/constants'
+import { kanjiToNumber } from '@/utils/japanese'
 
 /**
  * Translates Japanese rank text to English format.
@@ -42,11 +42,20 @@ export function translateRank(rankText: string): string {
 /**
  * Translates Japanese record text to English format.
  *
- * @param recordText - Japanese record text (e.g., "（6勝2敗）")
- * @returns English record format (e.g., "(6-2)")
+ * @param recordText - Japanese record text (e.g., "（6勝2敗）", "（1勝0敗3休）")
+ * @returns English record format (e.g., "(6-2)", "(1-0-3)")
  */
 export function translateRecord(recordText: string): string {
   if (!recordText) return ''
+
+  // Extract wins, losses, and rest days from pattern like "（6勝2敗）" or "（1勝0敗3休）"
+  const matchWithRest = recordText.match(/（(\d+)勝(\d+)敗(\d+)休）/)
+  if (matchWithRest) {
+    const wins = matchWithRest[1]
+    const losses = matchWithRest[2]
+    const rest = matchWithRest[3]
+    return `(${wins}-${losses}-${rest})`
+  }
 
   // Extract wins and losses from pattern like "（6勝2敗）"
   const match = recordText.match(/（(\d+)勝(\d+)敗）/)
